# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MagicAI  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/support/": {
            "get": {
                "tags": [
                    "Support"
                ],
                "summary": "Gets all support requests",
                "description": "Gets all support requests",
                "operationId": "supportRequests",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
        "/api/support/new-ticket": {
            "post": {
                "tags": [
                    "Support"
                ],
                "summary": "Create new support request",
                "description": "Create new support request",
                "operationId": "newTicket",
                "requestBody": {
                    "description": "Request support ticket data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "priority": {
                                        "description": "Priority",
                                        "type": "string",
                                        "enum": [
                                            "Low",
                                            "Normal",
                                            "High",
                                            "Critical"
                                        ]
                                    },
                                    "category": {
                                        "description": "Category",
                                        "type": "string",
                                        "enum": [
                                            "General Inquiry",
                                            "Technical Issue",
                                            "Improvement Idea",
                                            "Feedback",
                                            "Other"
                                        ]
                                    },
                                    "subject": {
                                        "description": "Subject of the Message",
                                        "type": "string"
                                    },
                                    "message": {
                                        "description": "Message to send",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}